monitor-server/src/main/java/com/devops/tools/server/service/graphql/exception/InternalServerException.java
package com.devops.tools.server.service.graphql.exception;

import com.fasterxml.jackson.annotation.JsonIgnore;
import com.google.common.annotations.VisibleForTesting;
import graphql.ErrorType;
import graphql.GraphQLError;
import graphql.language.SourceLocation;
import lombok.AllArgsConstructor;

import java.util.Collections;
import java.util.List;
import java.util.Map;

@AllArgsConstructor
public class InternalServerException extends RuntimeException implements GraphQLError {

    @VisibleForTesting
    static final String MESSAGE = "An Internal Server error occurred";

    @Override
    public String getMessage() {
        return MESSAGE;
    }

    @Override
    public List<SourceLocation> getLocations() {
        return null;
    }

    @Override
    public ErrorType getErrorType() {
        return ErrorType.DataFetchingException;
    }

    @Override
    public Map<String, Object> getExtensions() {
        return Collections.emptyMap();
    }

    @Override
    @JsonIgnore
    public StackTraceElement[] getStackTrace() {
        return super.getStackTrace();
    }

}